change pdb_serializer to download pdb files present in kegg.py output and to save list of actualy downloaded pdb files (this might differ - there are files in RCSB with no pdb files)

+++ pdb_serializer.py
@@ -1,29 +1,36 @@
+
 import pickle 
-import requests
+import argparse
 
+PDB_PATH = "pdb"
 
+import os, prody
 
-r = requests.get("http://www.genome.jp/dbget-bin/www_bget?hsa:10395")
-page = r.text
+if not os.path.exists(PDB_PATH):
+    os.mkdir(PDB_PATH)
 
-count = 0
-word = '1111' 
-ok = 0
-with open('set_pdb.pickle', 'rb') as f:
-	set_pdb = pickle.load(f)
-	for s in page:
-		
-		if (ok != 0):
-			if (ok < 5):
-				ok += 1
-			else:
-				set_pdb.add(word.upper())
-				ok = 0
-		word = word[1:]
-		word += s
-		l = len(word)
-		if (word == 'pdb:'):
-			ok = 1
+prody.proteins.localpdb.pathPDBFolder(PDB_PATH)
+
+def downloadPDBStructures(pdbIds):
+    """
+	download all structures if they are not present and were not previously processed 
+	returns list of structures which were actually downloaded
+	"""
+    downloadedPDBs = zip(pdbIds, prody.proteins.localpdb.fetchPDB(pdbIds)) 
+    downloadedPDBs = map(lambda x: x[0], filter(lambda x: x[1] is not None, downloadedPDBs)) 
+    return downloadedPDBs
+	
+if __name__ == "__main__":
+	gene_info_file = "gene_name_to_gene_ids_pdb_ids.pickle"
+	if not os.path.exists(gene_info_file):
+		print("File %s couldn't be found" % gene_info_file)
+		exit(1)
+	data = pickle.load(open(gene_info_file,"rb"))
+	# data.values() - contains (geneIds, pdbIds)
+	set_pdb = set()
+	for (_, pdbIds) in data.values():
+		set_pdb |= set(downloadPDBStructures(pdbIds))
+	# just in case we want to save the results
 
 	with open('set_pdb.pickle', 'wb') as f:
 		pickle.dump(set_pdb, f)
